feat: data rendering and streaming implemented
