+++ app/lib/data.ts
@@ -1,4 +1,3 @@
-import { sql } from "@vercel/postgres";
 import {
   CustomerField,
   CustomersTableType,
